+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.6]
+## Added
+- A workflow report using ezcharts.
+## Chnaged
+- Updates for the new Labs Launcher
+
 ## [v0.1.5]
 ### Changed
 - Replace samlmon for minimap2 for assigning reads to transcripts.